+++ src/main/java/org/kenuki/landingserver/repositories/LandingTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LandingTypeRepository extends JpaRepository<LandingType, Long> {
+public interface LandingTypeRepository extends JpaRepository<LandingType, Integer> {
 }
